// src/models/product.ts
export interface Unit {
	id: number;
	name: string;
}

export interface UnitSelect {
	value: number;
	label: string;
}

export interface Attribute {
	id: number;
	name: string;
	price: number;
}

export interface ProductAttribute {
	id: number;
	name: string;
	values: Attribute[];
}

export interface AttributeWithParentIdAndName extends Attribute {
	parentId: number;
	parentName: string;
}

export interface SearchHistory {
	id: number;
	keyword: string;
}

export interface PriceTable {
	uom_id: number;
	name: string;
	price_unit: number;
}

export interface Product {
	type: "product";
	id: number;
	id_product_att: number;
	name: string;
	barcode: string | boolean;
	product_available: number;
	star_rating: number;
	uom: Unit;
	uom_categ: Unit[];
	vat: number;
	description: string;
	description_sale?: string;
	price: number;
	price_orgin: number;
	price_list: {
		type: "percentage" | "fixed";
		min_quantity: number;
		value: number;
	}[];
	seller_price: number;
	sales_count: number;
	price_table: PriceTable[];
	attributes: ProductAttribute[];
	image_url: Array<string>;
	representative_image: Array<string>;
	variant_id?: number;
	company_id: number;
	company_name: string;
	price_origin: number;
	price_discount: number;
	categ_name: string;
	categ_id: number;
	rating_count: number;
	comment_count: number;
	wishlist: boolean;
	price_words: string;
}

export interface ProductDetailRes {
	id: number;
	product_tmpl_id: number;
	name: string;
	type: "product";
	uom: Unit;
	uom_categ: Unit[];
	company_id: number;
	company_name: string;
	description_sale: string;
	wholesales: Array<string>;
	categ_id: number;
	barcode: string;
	product_available: number;
	price: number;
	seller_price: number;
	image_url: Array<string>;
	qty_available: number;
	category: ProductDetailCategory;
	star_rating: 0.0;
}

export type ProductDetail = ProductDetailRes & Product;

export interface ProductDetailCategory {
	id: number;
	name: string;
}